+++ config.js
@@ -1,14 +1,19 @@
+// Configuración del portafolio
 const configuracion = {
     nombreEquipo: "Equipo [nombre]",
     version: "1.0.0",
-    fechaCreacion: "2025-11-20",
+    fechaCreacion: "2024-11-14",
     miembros: [
         {
-            nombre: "Francis",
-            rol: "Trabaje en proyecttos.html",
-            github: "Flancheesee"
+       nombre: "Alejandro Rodríguez Fernández",
+       rol: "Diseñador de proyectos",
+       github: "alejandrovsc20",
+        },
+        {
+       nombre: "Francisco Martin Jeronimo",
+       rol: "Coordinador",
+       github: "flancheesee"
         }
-        // Otros miembros añadirán sus datos aquí
     ]
 };